Change RabbitMQService.ts to interface, changed original implementation to AmqplibImpl.ts and added an implementation AmqpClientJSImpl.ts which use cloudamqp/amqp-client.js
