+++ package.json
@@ -16,6 +16,7 @@
    "author": "Everett Summer",
    "license": "Apache-2.0",
    "dependencies": {
+      "@cloudamqp/amqp-client": "^2.0.3",
       "@mikro-orm/core": "^5.1.3",
       "@mikro-orm/postgresql": "^5.1.3",
       "@sentry/node": "^6.19.2",
